+++ be/issue-tracker/.gitignore
@@ -36,4 +36,5 @@ out/
 ### VS Code ###
 .vscode/
 
+src/main/resources/application.properties
