+++ be/issue-tracker/src/main/java/com/issuetracker/member/service/LoginService.java
@@ -2,8 +2,8 @@ package com.issuetracker.member.service;
 
 import com.issuetracker.member.dto.LoginMemberDto;
 import com.issuetracker.member.dto.LoginTryDto;
+import com.issuetracker.member.entity.Member;
 import com.issuetracker.member.exception.LoginFailException;
-import com.issuetracker.member.model.Member;
 import com.issuetracker.member.repository.MemberRepository;
 import java.util.Optional;
 import lombok.RequiredArgsConstructor;
