--- be/issue-tracker/src/main/resources/application.properties
@@ -1,15 +1,0 @@
-#datasource
-spring.application.name=issue-tracker
-spring.datasource.url=${DB_URL}
-spring.datasource.username=${DB_USERNAME}
-spring.datasource.password=${DB_PASSWORD}
-#s3
-cloud.aws.credentials.access-Key=${S3_ACCESS_KEY}
-cloud.aws.credentials.secret-Key=${S3_SECRET_KEY}
-cloud.aws.region.static=ap-northeast-2
-cloud.aws.s3.bucket=sangchu-bucket
-cloud.aws.s3.dir=issue-tracker/img/
-#file size
-spring.servlet.multipart.max-file-size=10MB
-spring.servlet.multipart.max-request-size=10MB
-
